[Feature] exports utility functions in CronController
